requirements.txt
gradio>=4.0.0
rdflib
huggingface_hub
PyPDF2
python-docx
pandas
networkx
matplotlib
plotly
numpy
fastapi
uvicorn[standard]
pydantic
python-multipart
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
bitsandbytes>=0.41.0